refactor: organize data utility components into components/ directory
## Major Reorganization
Move 3 data utility components to proper location for better organization:
- gap_classifier.py → components/gap_classifier.py
- data_quality_validator.py → components/data_quality_validator.py
- timeframe_synchronizer.py → components/timeframe_synchronizer.py

## Rationale
These are pure utility components that process/analyze data but don't manage
data sources or implement core interfaces. They belong with other components
like GapAnalyzer, ProgressManager, and DataHealthChecker.

## Changes Made
- Move 3 utility files to components/ directory
- Update all import references across codebase (12 files updated)
- Move corresponding test files to tests/unit/data/components/
- Update components/__init__.py to export new components
- Fix import ordering with Ruff

## Final Organization
- **data/**: Core managers, adapters, interfaces, constants
- **data/components/**: 6 reusable utility components
  - DataHealthChecker, DataQualityValidator, GapAnalyzer
  - GapClassifier, ProgressManager, TimeframeSynchronizer

## Verification
- All 1078 unit tests pass
- Quality checks pass (lint, format, typecheck)
- Zero functional impact - all existing functionality intact
