+++ ktrdr/data/components/__init__.py
@@ -1,7 +1,18 @@
 """Data components module."""
 
 from .data_health_checker import DataHealthChecker
+from .data_quality_validator import DataQualityValidator
 from .gap_analyzer import GapAnalyzer
+from .gap_classifier import GapClassifier
 from .progress_manager import ProgressManager, ProgressState
+from .timeframe_synchronizer import TimeframeSynchronizer
 
-__all__ = ["DataHealthChecker", "GapAnalyzer", "ProgressManager", "ProgressState"]
+__all__ = [
+    "DataHealthChecker",
+    "DataQualityValidator",
+    "GapAnalyzer",
+    "GapClassifier",
+    "ProgressManager",
+    "ProgressState",
+    "TimeframeSynchronizer",
+]
