+++ ktrdr/data/components/data_health_checker.py
@@ -8,8 +8,8 @@ components, extracted to keep DataManager lean and maintain separation of concer
 import os
 from typing import Any, Optional
 
-from ktrdr.data.data_quality_validator import DataQualityValidator
-from ktrdr.data.gap_classifier import GapClassifier
+from ktrdr.data.components.data_quality_validator import DataQualityValidator
+from ktrdr.data.components.gap_classifier import GapClassifier
 from ktrdr.data.ib_data_adapter import IbDataAdapter
 from ktrdr.data.local_data_loader import LocalDataLoader
 from ktrdr.logging import get_logger
