+++ ktrdr/data/components/data_quality_validator.py
@@ -541,7 +541,10 @@ class DataQualityValidator:
             return
 
         # Use the existing gap classifier for intelligent detection
-        from ktrdr.data.gap_classifier import GapClassification, GapClassifier
+        from ktrdr.data.components.gap_classifier import (
+            GapClassification,
+            GapClassifier,
+        )
 
         gap_classifier = GapClassifier()
