+++ ktrdr/data/components/gap_analyzer.py
@@ -18,8 +18,8 @@ from typing import Any, Optional
 
 import pandas as pd
 
+from ktrdr.data.components.gap_classifier import GapClassification, GapClassifier
 from ktrdr.data.components.progress_manager import ProgressManager
-from ktrdr.data.gap_classifier import GapClassification, GapClassifier
 from ktrdr.data.loading_modes import DataLoadingMode
 from ktrdr.data.timeframe_constants import TimeframeConstants
 from ktrdr.logging import get_logger
