+++ ktrdr/data/components/timeframe_synchronizer.py
