+++ tests/unit/data/test_gap_analyzer.py
@@ -12,7 +12,11 @@ import pandas as pd
 import pytest
 
 from ktrdr.data.components.gap_analyzer import GapAnalyzer
-from ktrdr.data.gap_classifier import GapClassification, GapClassifier, GapInfo
+from ktrdr.data.components.gap_classifier import (
+    GapClassification,
+    GapClassifier,
+    GapInfo,
+)
 
 
 class TestGapAnalyzer:
