+++ tests/validation/test_gap_analysis_comprehensive.py
@@ -14,7 +14,11 @@ import pandas as pd
 import pytest
 
 from ktrdr.data.components.gap_analyzer import GapAnalyzer
-from ktrdr.data.gap_classifier import GapClassification, GapClassifier, GapInfo
+from ktrdr.data.components.gap_classifier import (
+    GapClassification,
+    GapClassifier,
+    GapInfo,
+)
 
 
 class TestModeSpecificGapAnalysis:
